+++ optimus/lib/optimus.dart
@@ -54,7 +54,8 @@ export 'src/form/checkbox_form_field.dart';
 export 'src/form/date_input_field.dart';
 export 'src/form/date_input_form_field.dart';
 export 'src/form/date_time_form_field.dart';
-export 'src/form/drawer_select_input.dart';
+export 'src/form/drawer/drawer_multiselect_input.dart';
+export 'src/form/drawer/drawer_select_input.dart';
 export 'src/form/error_variant.dart';
 export 'src/form/input_field.dart';
 export 'src/form/input_form_field.dart';
