+++ optimus/lib/src/common/group.dart
@@ -1,0 +1,4 @@
+import 'package:flutter/widgets.dart';
+
+typedef Grouper<T> = String Function(T item);
+typedef GroupBuilder = Widget Function(String value);
