+++ optimus/lib/src/common/value_builder.dart
@@ -1,0 +1,1 @@
+typedef ValueBuilder<T> = String Function(T value);
