+++ optimus/lib/src/form/drawer/common.dart
@@ -1,0 +1,4 @@
+import 'package:optimus/src/dropdown/dropdown_tile.dart';
+
+typedef OptimusDrawerListBuilder<T> =
+    List<OptimusDropdownTile<T>> Function(String query);
